requirements.txt
pandas
numpy
google-cloud-storage
google-cloud-aiplatform
# --- serving deps ---
fastapi
uvicorn
joblib
scikit-learn
